[analysis] separate the alias analysis into an equivalence analysis
The in-tree "alias analysis" was always a must-alias / pointer-equivalence
analysis wearing the mlir::AliasAnalysis costume: it never returns NoAlias,
its getModRef is a ModAndRef stub, and its load rule (values loaded from
must-alias refs are MustAlias) is only sound under Reussir's immutability
of shared box content -- a precondition invisible to a generic aggregate
client. The spill rule even relates loads through *distinct* stack slots
holding the same bytes: value equivalence, not address aliasing.

Make the contract honest:
- rename Analysis/AliasAnalysis.{h,cpp} to EquivalenceAnalysis.{h,cpp} and
  expose `EquivalenceAnalysis::areEquivalent(Value, Value) -> bool`
  ("same runtime bits", where false means unknown, never distinct), with
  the immutability precondition and the spill rule documented at the
  definition;
- keep registerAliasAnalysisImplementations as a thin adapter mapping
  equivalence to MustAlias so IncDecCancellation and RcDispatchFusion --
  the passes that need NoAlias from LocalAliasAnalysis -- keep their
  aggregate unchanged;
- switch InferVariantTag and TokenReuse, which only ever consumed
  MustAlias, to the direct equivalence API, dropping the per-pass
  aggregate construction;
- factor the duplicated projection walk into stripMatchingProjections,
  type-guard the value-to-index cache against recycled Values from
  erase-and-recreate churn (residual same-type case documented), and fix
  the `hueristic` typo.
