+++ lib/Analysis/CMakeLists.txt
@@ -1,7 +1,7 @@
 get_property(dialect_libs GLOBAL PROPERTY MLIR_DIALECT_LIBS)
 
 add_mlir_library(MLIRReussirAnalysis
-  AliasAnalysis.cpp
+  EquivalenceAnalysis.cpp
   LINK_LIBS PUBLIC
     immer
     MLIRAnalysis
